requirements.txt
Flask
Flask-RESTful
flasgger
marshmallow
jsons
flask-cors
Flask-SQLAlchemy
Flask-Migrate==2.7.0
psycopg2-binary==2.8.3
python-json-logger
iteration-utilities